proposer/db-utils/pyproject.toml
[project]
name = "python-db-explorations"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = ["python-dotenv"]

[project.scripts]
query-proofs = "query_proofs.py"
